refactor(hermes): split recovery boundary tests, delegate non-regular .env to validator
